+++ scripts/web_build.sh
@@ -3,12 +3,18 @@ set -e
 
 cd $(git rev-parse --show-toplevel)
 
+if [ -z "$POC_WEB_APP_GOOGLE_OAUTH_CLIENT_ID" ]; then
+    echo "Missing env variable: POC_WEB_APP_GOOGLE_OAUTH_CLIENT_ID"
+    exit 1
+fi
+
 echo "web_build started"
 
 echo "Copying 'po_contacts_flutter' to 'bin/tmp'..."
 rm -rf bin/tmp
 rm -rf bin/web
 cp -r po_contacts_flutter bin/tmp
+echo "const POC_GOOGLE_OAUTH_CLIENT_ID = '$POC_WEB_APP_GOOGLE_OAUTH_CLIENT_ID';">bin/tmp/lib/assets/constants/google_oauth_client_id.dart
 
 cd bin/tmp
